separate out contributed coi
