+++ src/utils.js
@@ -306,6 +306,7 @@ export function spatial_abilities (id) {
         county_brush: true,
         shapefile: true,
         find_unpainted: true,
+        load_coi: true,
       },
       alaska: {
         number_markers: true,
@@ -402,6 +403,7 @@ export function spatial_abilities (id) {
         contiguity: 2,
         shapefile: true,
         find_unpainted: true,
+        load_coi: true,
       },
       connecticut: {
         county_brush: true,
@@ -410,6 +412,7 @@ export function spatial_abilities (id) {
         contiguity: 2,
         shapefile: true,
         find_unpainted: true,
+        load_coi: true,
       },
       dc: {
         multiyear: 2018,
@@ -506,6 +509,7 @@ export function spatial_abilities (id) {
         county_brush: true,
         shapefile: true,
         find_unpainted: true,
+        load_coi: true,
       },
       iowa: {
         number_markers: true,
@@ -545,6 +549,7 @@ export function spatial_abilities (id) {
         contiguity: 2,
         shapefile: true,
         find_unpainted: true,
+        load_coi: true,
       },
         la_vra: {
           native_american: true,
@@ -570,6 +575,7 @@ export function spatial_abilities (id) {
         absentee: true,
         shapefile: true,
         find_unpainted: true,
+        load_coi: true,
       },
         baltimore: {
           border: true,
@@ -648,6 +654,7 @@ export function spatial_abilities (id) {
         number_markers: true,
         county_brush: true,
         shapefile: true,
+        load_coi: true,
         // find_unpainted: true,
       },
       montana: {
@@ -665,6 +672,7 @@ export function spatial_abilities (id) {
         county_brush: true,
         absentee: true,
         shapefile: true,
+        load_coi: true,
         // find_unpainted: true,
       },
       nevada: {
@@ -864,6 +872,7 @@ export function spatial_abilities (id) {
         shapefile: true,
         find_unpainted: true,
         sideload: true,
+        load_coi: true,
       },
         philadelphia: {
           number_markers: true,
@@ -960,6 +969,7 @@ export function spatial_abilities (id) {
         // native_american: true,
         shapefile: true,
         // find_unpainted: true,
+        load_coi: true,
       },
         vabeach: {
           multiyear: 2018,
